fix(lint): move slash resolver to graph/ so operator_api doesn't import server
The shared slash-command resolver landed in server.chat, but the import-linter
contract forbids operator_api -> server, so the palette importing it broke the
gate. Moved _slash_kind / find_user_facing_skill / resolve_slash_commands +
slugify into a neutral graph/slash_commands.py that both layers may import
(operator_api -> graph and server -> graph are both allowed). Behavior unchanged;
dropped the now-unused `import re` from server.chat.
